# golang/invest-tracker/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for tracking and analyzing investments across different asset types",
        "title": "Investment Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.systentando.com.br/support",
            "email": "support@systentando.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7777",
    "basePath": "/api/v1",
    "paths": {
...
        "/simulations": {
            "get": {
                "description": "Get a list of all simulations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Get all simulations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "pending",
                            "running",
                            "completed",
                            "failed"
                        ],
                        "type": "string",
                        "description": "Simulation status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of simulations",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new investment simulation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Create new simulation",
                "parameters": [
                    {
                        "description": "Simulation data",
                        "name": "simulation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created simulation",
                        "schema": {
                            "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/simulations/run": {
            "post": {
                "description": "Execute a simulation with given parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Run simulation",
                "parameters": [
                    {
                        "description": "Simulation parameters",
                        "name": "simulation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Simulation execution results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/simulations/{id}": {
            "get": {
                "description": "Get a specific simulation by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Get simulation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Simulation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Simulation details",
                        "schema": {
                            "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation"
                        }
                    },
                    "404": {
                        "description": "Simulation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing simulation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Update simulation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Simulation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Simulation data",
                        "name": "simulation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated simulation",
                        "schema": {
                            "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Simulation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a simulation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Delete simulation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Simulation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Simulation deleted successfully"
                    },
                    "404": {
                        "description": "Simulation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/simulations/{id}/results": {
            "get": {
                "description": "Get results of a specific simulation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "simulations"
                ],
                "summary": "Get simulation results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Simulation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Simulation results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Simulation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_analysis_entity.InvestmentStrategy": {
            "type": "string",
            "enum": [
                "value",
                "growth",
                "dividend",
                "trend",
                "momentum",
                "dca",
                "swing",
                "diversification"
            ],
            "x-enum-comments": {
                "StrategyDCA": "Dollar-cost averaging",
                "StrategyDiversification": "Diversificação",
                "StrategyDividend": "Investimento em dividendos",
                "StrategyGrowth": "Investimento em crescimento",
                "StrategyMomentum": "Análise de momentum",
                "StrategySwingTrading": "Swing trading",
                "StrategyTrend": "Análise de tendências",
                "StrategyValue": "Investimento em valor"
            },
            "x-enum-descriptions": [
                "Investimento em valor",
                "Investimento em crescimento",
                "Investimento em dividendos",
                "Análise de tendências",
                "Análise de momentum",
                "Dollar-cost averaging",
                "Swing trading",
                "Diversificação"
            ],
            "x-enum-varnames": [
                "StrategyValue",
                "StrategyGrowth",
                "StrategyDividend",
                "StrategyTrend",
                "StrategyMomentum",
                "StrategyDCA",
                "StrategySwingTrading",
                "StrategyDiversification"
            ]
        },
...
        "github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.Simulation": {
            "type": "object",
            "properties": {
                "asset": {},
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "entryPrice": {
                    "type": "number"
                },
                "errorMessage": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "initialAmount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.SimulationResult"
                    }
                },
                "scenarios": {
                    "description": "Variações percentuais",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "status": {
                    "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.SimulationStatus"
                },
                "stopLoss": {
                    "type": "number"
                },
                "strategy": {
                    "$ref": "#/definitions/github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_analysis_entity.InvestmentStrategy"
                },
                "targetPrice": {
                    "type": "number"
                },
                "timeHorizon": {
                    "description": "Em dias",
                    "type": "integer"
                }
            }
        },
        "github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.SimulationResult": {
            "type": "object",
            "properties": {
                "annualizedROI": {
                    "type": "number"
                },
                "finalAmount": {
                    "type": "number"
                },
                "finalPrice": {
                    "type": "number"
                },
                "profitLoss": {
                    "type": "number"
                },
                "profitLossPerc": {
                    "type": "number"
                },
                "scenarioChange": {
                    "description": "Percentual",
                    "type": "number"
                }
            }
        },
        "github_com_systentandobr_backend-monorepo_golang_invest-tracker_internal_domain_simulation_entity.SimulationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "complete",
                "failed"
            ],
            "x-enum-comments": {
                "SimulationComplete": "Simulação concluída",
                "SimulationFailed": "Simulação falhou",
                "SimulationPending": "Simulação pendente",
                "SimulationRunning": "Simulação em execução"
            },
            "x-enum-descriptions": [
                "Simulação pendente",
                "Simulação em execução",
                "Simulação concluída",
                "Simulação falhou"
            ],
            "x-enum-varnames": [
                "SimulationPending",
                "SimulationRunning",
                "SimulationComplete",
                "SimulationFailed"
            ]
        }
    }
}