# golang/invest-tracker/internal/adapter/controller/simulation_controller.go
package controller

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/systentandobr/backend-monorepo/golang/invest-tracker/internal/domain/simulation/entity"
	"github.com/systentandobr/backend-monorepo/golang/invest-tracker/pkg/common/logger"
)

// SimulationController handles simulation-related HTTP requests
type SimulationController struct {
	logger logger.Logger
}

// NewSimulationController creates a new simulation controller
func NewSimulationController(logger logger.Logger) *SimulationController {
	return &SimulationController{
		logger: logger,
	}
}

// RegisterRoutes registers simulation routes
func (c *SimulationController) RegisterRoutes(router *gin.RouterGroup) {
	simulations := router.Group("/simulations")
	{
		simulations.GET("", c.GetSimulations)
		simulations.GET("/:id", c.GetSimulationByID)
		simulations.POST("", c.CreateSimulation)
		simulations.PUT("/:id", c.UpdateSimulation)
		simulations.DELETE("/:id", c.DeleteSimulation)
		simulations.POST("/run", c.RunSimulation)
		simulations.GET("/:id/results", c.GetSimulationResults)
	}
}

// GetSimulations godoc
// @Summary Get all simulations
// @Description Get a list of all simulations
// @Tags simulations
// @Accept json
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param status query string false "Simulation status" Enums(pending, running, completed, failed)
// @Success 200 {object} map[string]interface{} "List of simulations"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations [get]
func (c *SimulationController) GetSimulations(ctx *gin.Context) {
	page := ctx.DefaultQuery("page", "1")
	limit := ctx.DefaultQuery("limit", "10")
	status := ctx.Query("status")
	// TODO: Implement get simulations logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get simulations endpoint",
		"page":   page,
		"limit":  limit,
		"status": status,
	})
}

// GetSimulationByID godoc
// @Summary Get simulation by ID
// @Description Get a specific simulation by its ID
// @Tags simulations
// @Accept json
// @Produce json
// @Param id path string true "Simulation ID"
// @Success 200 {object} entity.Simulation "Simulation details"
// @Failure 404 {object} map[string]interface{} "Simulation not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations/{id} [get]
func (c *SimulationController) GetSimulationByID(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement get simulation by ID logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get simulation by ID endpoint",
		"id":      id,
	})
}

// CreateSimulation godoc
// @Summary Create new simulation
// @Description Create a new investment simulation
// @Tags simulations
// @Accept json
// @Produce json
// @Param simulation body entity.Simulation true "Simulation data"
// @Success 201 {object} entity.Simulation "Created simulation"
// @Failure 400 {object} map[string]interface{} "Bad request"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations [post]
func (c *SimulationController) CreateSimulation(ctx *gin.Context) {
	var simulation entity.Simulation
	if err := ctx.ShouldBindJSON(&simulation); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement create simulation logic
	ctx.JSON(http.StatusCreated, gin.H{
		"message":     "Create simulation endpoint",
		"simulation": simulation,
	})
}

// UpdateSimulation godoc
// @Summary Update simulation
// @Description Update an existing simulation
// @Tags simulations
// @Accept json
// @Produce json
// @Param id path string true "Simulation ID"
// @Param simulation body entity.Simulation true "Simulation data"
// @Success 200 {object} entity.Simulation "Updated simulation"
// @Failure 400 {object} map[string]interface{} "Bad request"
// @Failure 404 {object} map[string]interface{} "Simulation not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations/{id} [put]
func (c *SimulationController) UpdateSimulation(ctx *gin.Context) {
	id := ctx.Param("id")
	var simulation entity.Simulation
	if err := ctx.ShouldBindJSON(&simulation); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement update simulation logic
	ctx.JSON(http.StatusOK, gin.H{
		"message":     "Update simulation endpoint",
		"id":          id,
		"simulation": simulation,
	})
}

// DeleteSimulation godoc
// @Summary Delete simulation
// @Description Delete a simulation by ID
// @Tags simulations
// @Accept json
// @Produce json
// @Param id path string true "Simulation ID"
// @Success 204 "Simulation deleted successfully"
// @Failure 404 {object} map[string]interface{} "Simulation not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations/{id} [delete]
func (c *SimulationController) DeleteSimulation(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement delete simulation logic
	ctx.JSON(http.StatusNoContent, gin.H{
		"message": "Delete simulation endpoint",
		"id":      id,
	})
}

// RunSimulation godoc
// @Summary Run simulation
// @Description Execute a simulation with given parameters
// @Tags simulations
// @Accept json
// @Produce json
// @Param simulation body map[string]interface{} true "Simulation parameters"
// @Success 200 {object} map[string]interface{} "Simulation execution results"
// @Failure 400 {object} map[string]interface{} "Bad request"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations/run [post]
func (c *SimulationController) RunSimulation(ctx *gin.Context) {
	var params map[string]interface{}
	if err := ctx.ShouldBindJSON(&params); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement run simulation logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Run simulation endpoint",
		"params":  params,
	})
}

// GetSimulationResults godoc
// @Summary Get simulation results
// @Description Get results of a specific simulation
// @Tags simulations
// @Accept json
// @Produce json
// @Param id path string true "Simulation ID"
// @Success 200 {object} map[string]interface{} "Simulation results"
// @Failure 404 {object} map[string]interface{} "Simulation not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /simulations/{id}/results [get]
func (c *SimulationController) GetSimulationResults(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement get simulation results logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get simulation results endpoint",
		"id":      id,
	})
}